requirements.txt
flask
flask-sqlalchemy
pyjwt
werkzeug
pymysql
mysql-connector-python